jademirror/backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
requests==2.32.3
openai==1.76.2
